add register page, and forward data for payment gateway using PHP
A registration, with a form, has been added. The form data will be
sent to a PHP page. The PHP page will send the necessary data (e.g.,
customer info, amount, type, etc.) to SSLCOMMERZ and then, if
receives green signal, forwards to the SSLCOMMERZ page for selecting
payment method.